aarch64: Attempt to exit via QEMU on aarch64
This will help with testing (rusty-loader) since we currently
can't go very far in the kernel.
However if we reach this point, then that is already useful
for testing rusty-loader in CI.
This requires passing the `-semihosting` option to QEMU.
If this is not done or real hardware is used, a spinloop
will be entered.